requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
arxiv==1.4.8
openai==1.12.0
python-dotenv==1.0.0
schedule==1.2.1
pandas==2.1.4
numpy==1.24.3 